+++ lib/src/core/better_player_controller.dart
@@ -769,7 +769,6 @@ class BetterPlayerController extends ChangeNotifier {
     _controlsVisibilityStreamController.add(controlsAlwaysVisible);
   }
 
-
   ///Dispose BetterPlayerController. When [forceDispose] parameter is true, then
   ///autoDispose parameter will be overridden and controller will be disposed
   ///(if it wasn't disposed before).