+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.48
 * Fixed loading large videos in iOS.
 * Fixed partly progress bar jumping when seek issue in iOS.
+* Added forceDispose parameter to dispose method in BetterPlayerController.
 
 ## 0.0.47
 * Fixed Android loading indicator issue.